Stop calling rosdep directly, allow a command to be specified
This command will be placed directly into the Dockerfile, instead of
embedding an apt install with the result of running rosdep. This was
done because calling rosdep was mildly gross, and because it puts a hard
limit on it existing in the enviroment. We will be pulling docker cache
layers anyway, and locally these will not be invalidated (instead this
program would have run again) so this is an improvement