Use only BTreeMap
Originally a choice was made to use both HashMap and BTreeMap, because a
HashMap is more efficient for building up the popularity, but a BTreeMap
was needed later for ordered data. A copy is required to go from
`Map<Name, u32>` to `Map<u32, Vec<Name>>` anyway, so why not change
datastructure. However, there are now multiple places that would benefit
in readability from using the same datastructure, and the slight
efficiency is not worth the readability hit