A single layer is generated at a time, fully resolving dependencies
To enable the abstraction of generating both build layers, which only
use build dependencies, and exec layers, which use run time
dependencies, it was best to switch to generating a full layer at a
time, rather than a set of layers all together, then process all those
layers to resolve the system dependencies.
This meant refactoring the `generate_layers` function to
`generate_layer`.
This allows the function to take in a list of dependencies without
knowing which set they came from